scientific_article_analyzer/requirements.txt
# Scientific Article Analysis System - Dependencies

# Core Framework
fastmcp==0.7.0
pydantic==2.5.0
python-dotenv==1.0.0

# Machine Learning & AI
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# Text Processing
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==1.1.0
markdown==3.5.1

# Data Handling
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Web Framework (for MCP Server)
uvicorn==0.24.0
fastapi==0.104.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Optional: Enhanced AI Capabilities
openai==1.3.0
anthropic==0.7.0

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.25.2

# Model Context Protocol
mcp==1.0.0

# System & Utilities
rich==13.6.0
typer==0.9.0
tqdm==4.66.1
click==8.1.7
loguru==0.7.2

# Configuration & Validation
pyyaml==6.0.1
jsonschema==4.19.1